Remove DidDocument class.
